backend/requirements.txt
# ============================================
# URWA Brain - Requirements
# Production-Grade Web Intelligence Platform
# ============================================

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
slowapi>=0.1.9


# Web Scraping
playwright>=1.41.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
requests>=2.31.0
httpx>=0.26.0
aiohttp>=3.9.0               # Async HTTP for compliance checks

# LLM Integration
google-genai>=0.4.0
groq>=0.4.0

# Logging & Monitoring
loguru>=0.7.0
psutil>=5.9.0                # System metrics (CPU, memory, disk)

# Data Processing
pandas>=2.0.0
fpdf2>=2.7.0

# Search
ddgs>=1.0.0                  # DuckDuckGo Search (Valid package)
googlesearch-python>=1.2.0   # Official Google Search (No API)

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# ============================================
# OPTIONAL - Enhanced Features
# ============================================

# OCR for Visual Extraction (optional)
# pytesseract>=0.3.10
# Pillow>=10.0.0

# Database (optional - for production)
# redis>=5.0.0
# sqlalchemy>=2.0.0

# Proxy Management (optional)
# python-socks>=2.4.0

# ============================================
# CAPTCHA Solving (API key required)
# ============================================
# No pip packages needed - uses HTTP APIs:
# - 2Captcha: http://2captcha.com
# - Anti-Captcha: https://anti-captcha.com

# ============================================
# Install with:
# pip install -r requirements.txt
# playwright install chromium
# ============================================
